model_engine.py
import torch
from PIL import Image
from PIL.ExifTags import TAGS
import io
from transformers import BlipProcessor, BlipForConditionalGeneration

# 1. NEURAL CORE INITIALIZATION
# We use BLIP for 'everything' recognition to avoid the 'blindness' of CLIP.
device = "cuda" if torch.cuda.is_available() else "cpu"
processor = BlipProcessor.from_pretrained("Salesforce/blip-image-captioning-base")
model = BlipForConditionalGeneration.from_pretrained("Salesforce/blip-image-captioning-base").to(device)

def extract_metadata(image_bytes):
    """
    TRUTH AUDIT: This function looks for real camera data.
    If it's a web image with no EXIF, it returns 'N/A' to stay honest.
    """
    try:
        img = Image.open(io.BytesIO(image_bytes))
        exif_data = img._getexif()
        
        # We start with empty strings so the UI knows there is no real data
        metadata = {
            "Model": "Digital Source", 
            "Make": "Unknown", 
            "ISO": "N/A", 
            "F": "N/A"
        }
        
        if exif_data:
            for tag, value in exif_data.items():
                decoded = TAGS.get(tag, tag)
                if decoded == "Model": 
                    metadata["Model"] = str(value)
                elif decoded == "Make": 
                    metadata["Make"] = str(value)
                elif decoded == "ISOSpeedRatings": 
                    metadata["ISO"] = str(value)
                elif decoded == "FNumber": 
                    # Converting RAW metadata to readable f-stop
                    try:
                        metadata["F"] = f"f/{float(value)}"
                    except:
                        metadata["F"] = "N/A"
        return metadata
    except Exception:
        return {"Model": "N/A", "Make": "N/A", "ISO": "N/A", "F": "N/A"}

def predict_image(image_bytes):
    """
    The Brain: Recognizes objects and applies professional insights.
    """
    try:
        raw_image = Image.open(io.BytesIO(image_bytes)).convert('RGB')
        
        # Run Metadata Extraction first
        tech = extract_metadata(image_bytes)

        # GENERATIVE ANALYSIS (BLIP Transformer)
        inputs = processor(raw_image, return_tensors="pt").to(device)
        
        # repetition_penalty=2.5 is the 'fix' for your Giza 'GI GI' error.
        out = model.generate(
            **inputs, 
            max_new_tokens=35, 
            repetition_penalty=2.5, 
            num_beams=5
        )
        description = processor.decode(out[0], skip_special_tokens=True)

        # Format Title for VISIORYX UI
        title = description.upper()
        
        # DYNAMIC INSIGHT LOGIC
        # We use keywords from the AI's description to provide 'expert' advice.
        insight = "Neural Analysis: Scene successfully mapped into Visioryx ecosystem."
        
        if any(word in description for word in ["man", "woman", "person", "girl", "boy"]):
            insight = "Human Intel: Professional subject detected. Composition is scout-ready."
        elif any(word in description for word in ["tower", "building", "landmark", "pyramid", "castle"]):
            insight = "Geospatial Data: Architectural landmark identified. GPS tracking available."
        elif any(word in description for word in ["car", "vehicle", "truck"]):
            insight = "Industrial Intel: Automotive design detected. Evaluating aerodynamic profiles."
        elif any(word in description for word in ["tree", "flower", "animal", "bird"]):
            insight = "Biological Data: Natural species identified. Cross-referencing wildlife database."

        return {
            "title": title,
            "confidence": "VERIFIED MATCH",
            "device": f"{tech['Make']} {tech['Model']}",
            "iso": tech['ISO'],
            "aperture": tech['F'],
            "insight": insight
        }
    except Exception as e:
        return {
            "title": "NEURAL ERROR", 
            "insight": f"Analysis failed: {str(e)}", 
            "device": "System Error",
            "confidence": "0%",
            "iso": "N/A",
            "aperture": "N/A"
        }