requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
transformers
open_clip_torch